Renamed phonons folder to vibrational
